+++ cakeManage/admin.py
@@ -11,7 +11,3 @@ admin.site.register(AmountCoupon)
 admin.site.register(PercentCoupon)
 admin.site.register(Menu_Color)
 admin.site.register(Menu_Cream)
-# class ReviewAdmin(admin.ModelAdmin):
-#     list_display = ['user','referred_cake','rate', 'comment']
-#     readonly_fields = ['pub_date',]
-admin.site.register(Store_Menu)
